microbitmatrix: refactor to eliminate duplicate code with microbit v1/v2
